Update splash.js
I changed a lot in the code. I also started using some constructor function for some of the objects. I modified the splash screen to be more playful and included a "customize" button that allows the user to change the colors of the player's pants, shirt, and hair (which are global variables). I also added code to keep score and life count, as well as allow the player to move on the screen using the arrow keys. The only problem is I did not have enough time or patience to fix the collision area (something maybe you can do). I also included code so that when you run out of lives "currentScene = 4", so the game will "freeze" because I have not made a "gameOverScreen." There is an area in the code for this, I just ran out of time to complete this part today. I did not include too many comments (sorry about that), but I can do that later in the week, or explain the code in person. (I  might have missed some minor things, but yeah... I did a lot. XD)